requirements.txt
openai==0.10.1
streamlit==0.85.0
python-dotenv==0.20.0
